feat(3.0): Modern Alchemist — full 3.0 hardening sprint
Navigation & chrome
- Migrated all auth-gated + app-surface pages into (alchm) route group (dark shell):
  birth-chart, current-chart, recipe-generator, planetary-chart, restaurant-creator,
  commensal, feed, cosmic-recipe, generated-recipe, food-tracking, profile/*
- /commensal recolored from legacy purple → alchm copper/violet design tokens

Onboarding
- Added "Skip for now" CTA to /onboarding page (PATCH /api/onboarding {skipNatal:true})
- NatalPromptBanner on (alchm)/page.tsx — shows on ?prompt=natal, dismissable via localStorage

MenuPlannerContext refactor
- 2182-line monolith → 5 focused modules in src/contexts/menu-planner/:
  types.ts (244 lines), useWeekNavigation.ts (65), useMealSlots.ts (498),
  MenuPlannerProvider.tsx (1280). Barrel at MenuPlannerContext.tsx (28 lines).
  Public API unchanged — no consumer files touched.

Analytics
- track("command_palette_open") in CommandPalette.tsx
- track("upgrade_gate_shown"), track("upgrade_gate_converted") in AuthFollowups.tsx
- track("auth_handshake_completed") in AuthHandshake component

Documentation
- CHANGELOG.md — keep-a-changelog format from v1.0.0 → 3.0.0
- README.md — fully rewritten: v3.0, new architecture, project structure, env vars, ADR index
- docs/API_REFERENCE.md — every /api/* route documented (method, auth, schema)
- docs/adr/001–005 — Architecture Decision Records for nav IA, two-tier pricing,
  token economy throttle, device sessions, denormalized read_model

Release
- package.json: version 2.3.0 → 3.0.0

Build: 0 errors, 0 warnings. 45 static pages generated.
