+++ src/app/(alchm)/generated-recipe/[id]/page.tsx
