+++ src/app/(alchm)/generated-recipe/page.tsx
